- Refactor for Azure OpenAI - Moved Readme and refactored for Azure OpenAI - Renamed directories for better python module names
