+++ requirements.txt
@@ -1,0 +1,4 @@
+openai==1.69.0
+tiktoken==0.9.0
+pgvector==0.4.0
+psycopg2-binary==2.9.10
